# rpc : make RPC servers come first in the device list