- [x] I have read the [contributing guidelines](https://github.com/ggerganov/llama.cpp/blob/master/CONTRIBUTING.md)
- Self-reported review complexity:
  - [x] Low
  - [ ] Medium
  - [ ] High

This patch implements an idea suggested by @slaren which is to make RPC servers come first in the device list. When the last device is a local one, we don't have to transfer the logits over the network and this have significant impact on the performance.

Here are the results for my laptop (NVIDIA T1200) offloading to an RPC server with Tesla T4 on 100Mb/s network:

| GPU                     | Model           |   Model Size [GiB] |   Num. of Par. | Test   |   t/s master |   t/s rpc-first |   Speedup |
|:------------------------|:----------------|-------------------:|---------------:|:-------|-------------:|----------------:|----------:|
| NVIDIA T1200 Laptop GPU | llama 1B Q8_0   |               1.09 |     1100048384 | pp512  |      1742.68 |         1987.90 |      1.14 |
| NVIDIA T1200 Laptop GPU | llama 1B Q8_0   |               1.09 |     1100048384 | tg128  |        67.98 |           82.49 |      1.21 |
| NVIDIA T1200 Laptop GPU | llama 8B Q4_K_M |               4.58 |     8030261248 | pp512  |       356.96 |          465.93 |      1.31 |
| NVIDIA T1200 Laptop GPU | llama 8B Q4_K_M |               4.58 |     8030261248 | tg128  |        27.36 |           29.95 |      1.09 |

The amount of network traffic sent from the RPC server to the main host is from 4x to 15x less with this patch.