@awatuna 2024-09-04
I think this broke the Windows cuda llama-server's rpc.

The Windows cuda llama-server used to ignore --rpc.

However if you build it with GGML_RPC set then the Windows cuda llama-server, eg b3649 with #9184 patch, works utilizing both local Windows gpu and remote Linux gpu.

Now build with GGML_RPC set, the Windows cuda llama-server ignores --rpc again.