@rgerganov 2024-09-05
> Now build with GGML_RPC set, the Windows cuda llama-server ignores --rpc again.

Hm, I don't have a windows machine atm but I cannot reproduce this on Linux. What do you mean the `--rpc` option is ignored? Can you start `rpc-server` on localhost and paste the output of:

```
llama-server -m <some_model.gguf> --rpc localhost:50052 -ngl 99
```