@awatuna 2024-09-05
The official Windows cuda build
https://github.com/ggerganov/llama.cpp/releases/download/b3669/llama-b3669-bin-win-cuda-cu12.2.0-x64.zip

It does not build rpc-server, and the llama-server prints help when --rpc is set.
No rpc at all.

The merged patch from another collaborator that enables rpc by default,
only enabled rpc for Windows builds, not Windows cuda builds.
Windows builds and Windows cuda builds are in different parts of the CI workflow, the patch missed it.
So I made another patch, #9184 to enable it.

The resulting b3649 with GGML_RPC set for Windows cuda build
https://github.com/awatuna/llama.cpp/releases/tag/b3655

llama-server rpc works

llm_load_tensors: ggml ctx size =    0.68 MiB
llm_load_tensors: offloading 46 repeating layers to GPU
llm_load_tensors: offloading non-repeating layers to GPU
llm_load_tensors: offloaded 47/47 layers to GPU
llm_load_tensors: RPC[192.168.200.246:50052] buffer size =  9228.81 MiB
llm_load_tensors:        CPU buffer size =  1195.31 MiB
llm_load_tensors:      CUDA0 buffer size = 18362.25 MiB

However, b3664 with GGML_RPC set for Windows cuda build
https://github.com/awatuna/llama.cpp/releases/tag/b3671

llama-server accepts --rpc, but it didn't even try to connect to rpc server.

llm_load_tensors: ggml ctx size =    0.68 MiB
llama_model_load: error loading model: unable to allocate backend buffer
llama_load_model_from_file: failed to load model