@awatuna 2024-09-05
If the rpc server is down, win cuda b3649 with rpc will spam

Failed to connect to 192.168.200.246:50052

win cuda b3664 with rpc just ignores the --rpc, doesn't try to connect at all.